fix: moved dtos and mapper for clean architecture logic, reverted generic service
