+++ TTE.Application/DTOs/EmployeeResponseDto.cs
@@ -1,12 +1,16 @@
-﻿namespace TTE.Infrastructure.DTOs
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace TTE.Application.DTOs
 {
-    public class UserResponseDto
+    public class EmployeeResponseDto
     {
         public int Id { get; set; }
         public string UserName { get; set; } = string.Empty;
         public string Email { get; set; } = string.Empty;
-        public string Name { get; set; } = string.Empty;
-        public string Password { get; set; } = string.Empty;
         public string Role { get; set; } = string.Empty;
     }
 }
