+++ TTE.Application/DTOs/LoginRequestDto.cs
@@ -5,7 +5,7 @@ using System.Text;
 using System.Threading.Tasks;
 using TTE.Commons.Validators;
 
-namespace TTE.Infrastructure.DTOs
+namespace TTE.Application.DTOs
 {
     public class LoginRequestDto
     {
