+++ TTE.Application/DTOs/ProductPaginatedResponseDto.cs
@@ -1,4 +1,4 @@
-﻿namespace TTE.Infrastructure.DTOs
+﻿namespace TTE.Application.DTOs
 {
     public class ProductPaginatedResponseDto : GenericResponseDto<IEnumerable<ProductResponseDto>>
     {
