+++ TTE.Application/DTOs/UserResponseDto.cs
@@ -1,16 +1,12 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace TTE.Infrastructure.DTOs
+﻿namespace TTE.Application.DTOs
 {
-    public class EmployeeResponseDto
+    public class UserResponseDto
     {
         public int Id { get; set; }
         public string UserName { get; set; } = string.Empty;
         public string Email { get; set; } = string.Empty;
+        public string Name { get; set; } = string.Empty;
+        public string Password { get; set; } = string.Empty;
         public string Role { get; set; } = string.Empty;
     }
 }
